# perf: Improve local variable reference finding