This PR improves the time to find references and rename local variables. As suggested by @Techatrix in the discord, for certain cases the search for references can be limited to the current document rather than the entire project. After some brief testing with function-local variables in the zls project I saw a typical 4,500ms to <10ms improvement with debug builds. Testing with some of the larger files in zig itself, I saw a typical 20,000ms to 2,500ms improvement. Greater improvements are seen when using a `ReleaseSafe` build. I can spend some time gathering more numbers if that would be helpful :)

Related Issue: #1706