@WillLillis 2024-06-16
Took another pass over this and sped up private container-scoped decls as well. In my own testing this didn't break anything/ cause any refs to be missed, but would appreciate a second look-over.