@WillLillis 2024-06-16
> After taking a closer look at the references code, I realized that this extra logic should be moved [here](https://github.com/zigtools/zls/blob/b4f4d35354f39fd190b9ce681c75affafaa0a1ab/src/features/references.zig#L239) since its the only place where `workspace` is being used. This makes it clear that we only need to handle the case where the declaration is an `.ast_node`. Otherwise the logic looks fine.
> 
> The code looked a bit different from what I would have written so I left some suggestion on how to write it in a more "idiomatic" style. In the end, this is probably what I would have written:
> 
> ```zig
> const workspace = if (request == .highlight) false else blk: {
>     const doc_scope = try handle.getDocumentScope();
>     const scope_index = Analyser.innermostBlockScopeIndex(doc_scope, source_index).unwrap() orelse break :blk true;
>     break :blk switch (doc_scope.getScopeTag(scope_index)) {
>         .function, .block => false,
>         .container, .container_usingnamespace => decl.isPublic(),
>         .other => true,
>     };
> };
> ```
> 
> Hope that helps 🤷

That was really helpful, thank you! Getting this kind of feedback is always really appreciated so that so my cide can be cleaner moving forward :). I think I'll do another read through the language reference to brush up as well.